requirements.txt
aws-cdk-lib==2.178.1
constructs>=10.0.0,<11.0.0
pytest>=6.0
pytest-cov
# aws-cdk.aws-lambda
# aws-cdk.aws-apigateway
# boto3